feat: Add audio assets and begin Area2D player movement
This commit introduces new audio files to the project and refactors the player's movement logic.

The following audio files have been added and ignored in `.gitignore`:
- `assets/game-over-first-draft.wav`
- `assets/main_theme-first-draft.wav`
- `assets/victory-first-draft.wav`
Please download them from the mediafire like in the Assets/Readme.md .
It could take too long to git clone it hence this practice.

The Area2d is for the interaction between objects.
The `inputs` dictionary has been modified to store both the movement vector and an offset for the `Area2D`, this needs to be completed.
Additionally, the player_size variable has been introduced.